+++ apps/api/src/modules/branches.ts
@@ -1,6 +1,6 @@
 import { Elysia, status, t } from "elysia";
 
-import { slugFromString } from "@hebo/shared-api/utils/create-slug";
+import { slugFromString } from "@hebo/shared-api/utils/slug";
 
 import type { Prisma } from "~api/generated/prisma/client";
 import {
